large side panel update
