+++ NexTrace_Extension/react-app/src/components/Console.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function ConsoleComponent() {
+    
+    return (
+    <div>IM RETURNING ANOTHER REACT COMPONENT</div>
+    
+    )
+}
