--- NexTrace_Extension/react-app/src/components/PanelProvider.js
@@ -1,42 +1,0 @@
-export default function PanelProvider() {
-
-  return `
-    <!DOCTYPE html>
-    <html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Cat Coding</title>
-    </head>
-    <body>
-      <table>
-      <thead>
-        <tr>
-          <th>header1</th>
-          <th>header2</th>
-          <th>header3</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td>text1.1</td>
-          <td>text1.2</td>
-          <td>text1.3</td>
-        </tr>
-        <tr>
-          <td>text2.1</td>
-          <td>text2.2</td>
-          <td>text2.3</td>
-        </tr>
-        <tr>
-          <td>text3.1</td>
-          <td>text3.2</td>
-          <td>text3.3</td>
-        </tr>
-        <tr>
-        </tr>
-      </tbody>
-      </table>
-    </body>
-    </html>`;
-}
