+++ NexTrace_Extension/react-app/src/index.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import ReactDOM from 'react-dom';
-import App from './App.jsx';
+import App from './App';
+
 
 const root = createRoot(document.getElementById('root'));
 root.render(
-  <App />
+    <App/>
 );
-
-// ReactDOM.render(<App/>, document.getElementById('root'));
